--- logout.php
@@ -1,8 +1,0 @@
-<?php 
-    // LOGOUT ISLEMI
-	session_start();
-	session_destroy();
-
-	header("location:index.php");
-	exit;
-?>
